Move test into engine_test
