+++ app/btree/mock/tables.go
@@ -1,4 +1,4 @@
-package test_btree
+package mock
 
 import (
 	"encoding/binary"
